clean up of xeth
